Moved Twilio-sending code to CDTADPQ.data.notifications
