+++ CDTADPQ/data/notify.py
@@ -1,7 +1,7 @@
 import codecs, json, requests, uritemplate
 import os, psycopg2, psycopg2.extras
 
-
+from . import notifications
 from . import wildfires
 from . import zipcodes
 from . import users
@@ -11,14 +11,14 @@ def main():
     '''
     if os.environ['TWILIO_ACCOUNT'].startswith('AC'):
         # TODO move TwilioAccount out of users
-        twilio_account = users.TwilioAccount(
+        twilio_account = notifications.TwilioAccount(
             sid = os.environ.get('TWILIO_SID', ''),
             secret = os.environ.get('TWILIO_SECRET', ''),
             account = os.environ.get('TWILIO_ACCOUNT', ''),
             number = os.environ.get('TWILIO_NUMBER', '')
             )
     else:
-        twilio_account = users.TwilioAccount(
+        twilio_account = notifications.TwilioAccount(
             sid = codecs.decode(os.environ.get('TWILIO_SID', ''), 'rot13'),
             secret = codecs.decode(os.environ.get('TWILIO_SECRET', ''), 'rot13'),
             account = codecs.decode(os.environ.get('TWILIO_ACCOUNT', ''), 'rot13'),
@@ -94,17 +94,7 @@ TwilioURL = 'https://api.twilio.com/2010-04-01/Accounts/{account}/Messages.json'
 def send_notification(account, user, message):
     ''' Send fire notification to phone number
     '''
-    url = uritemplate.expand(TwilioURL, dict(account=account.account))
-    data = dict(From=account.number, To=user.phone_number, Body=message)
-    auth = account.sid, account.secret
-    posted = requests.post(url, auth=auth, data=data)
-    
-    if posted.status_code not in range(200, 299):
-        if 'message' in posted.json():
-            raise RuntimeError(posted.json()['message'])
-        else:
-            raise RuntimeError('Bad response from Twilio')
-    return True
+    return notifications.send_sms(account, user.phone_number, message)
 
 def log_user_notification(db, user, fire):
     ''' Log that the user has been notified of this emergency
